requirements.txt
fastapi
uvicorn
pydantic>=2.0.0
crewai
crewai_tools
sentence-transformers
psycopg2-binary
pgvector
langchain
langgraph
langchain-google-genai
langchain_community
tavily-python
python-dotenv
langchain_groq
thefuzz
langgraph-checkpoint-postgres
asyncio
asyncpg
psycopg_binary
google-api-python-client
langchain-openai
langsmith
celery
redis
gevent